+++ src/oci_cli/__init__.py
@@ -11,6 +11,7 @@ from .generated import blockstorage_cli  # noqa: F401
 from .generated import compute_cli  # noqa: F401
 from .generated import database_cli  # noqa: F401
 from .generated import dns_cli  # noqa: F401
+from .generated import filestorage_cli  # noqa: F401
 from .generated import identity_cli  # noqa: F401
 from .generated import loadbalancer_cli  # noqa: F401
 from .generated import objectstorage_cli  # noqa: F401
@@ -23,6 +24,7 @@ from . import database_cli_extended  # noqa: F401
 from . import dns_cli_extended  # noqa: F401
 from . import identity_cli_extended  # noqa: F401
 from . import objectstorage_cli_extended  # noqa: F401
+from . import filestorage_cli_extended  # noqa: F401
 from . import file_filters  # noqa: F401
 from . import final_command_processor  # noqa: F401
 from . import lb_cli_extended  # noqa: F401